+++ lib/internal/error_serdes.js
