+++ lib/internal/main/worker_thread.js
@@ -193,7 +193,7 @@ function workerOnGlobalUncaughtException(error, fromPromise) {
 
   let serialized;
   try {
-    const { serializeError } = require('internal/error-serdes');
+    const { serializeError } = require('internal/error_serdes');
     serialized = serializeError(error);
   } catch {}
   debug(`[${threadId}] uncaught exception serialized = ${!!serialized}`);
