+++ lib/internal/worker.js
@@ -42,7 +42,7 @@ const {
   ReadableWorkerStdio,
   WritableWorkerStdio
 } = workerIo;
-const { deserializeError } = require('internal/error-serdes');
+const { deserializeError } = require('internal/error_serdes');
 const { fileURLToPath, isURLInstance, pathToFileURL } = require('internal/url');
 
 const {
