+++ node.gyp
@@ -127,7 +127,7 @@
       'lib/internal/dtrace.js',
       'lib/internal/encoding.js',
       'lib/internal/errors.js',
-      'lib/internal/error-serdes.js',
+      'lib/internal/error_serdes.js',
       'lib/internal/fixed_queue.js',
       'lib/internal/freelist.js',
       'lib/internal/freeze_intrinsics.js',
